Trivial implementation.
I was not satisfied with current objc-property. Manually binding the
slots is good. But need some advanced post processing.

What about automatically generating? I don't think it might be a nice
idea too. The automatically interned property slots has the following
problem:
+ name might conflicts with lisp symbol if just using current
  `objc-intern' function
+ tooooooooooooooo many slots

But that might not be a problem, I will try to implement it in the
later.